FIX: nasty bug that xaxis and yaxis from redux store failing plotly js
This is a very nasty bug! When injecting data using redux store, if data
contains xaxis and yaxis props, the plot just failed silently. I can
only discover the bug when I switched to a redux wrapper approach (i.e.
useRef wrapper, making use of Plotly.react function to create the plot
instead, then chrome finally prompt for an error trace, saying that
"cannot assign to read only property 'xaxis' of '#<object>'", and after
gooling that is becuase plotly is trying to write to the 'data' object,
which is indeed not a plain js object, but a redux store object. The fix
is to creat a deep copy of the 'data' object before giving it to plotly.
My current solution is to use JSON.parse(JSON.stringify(data)).